requirements.txt
langchain
langchain-community
python-dotenv
streamlit
unstructured
tiktoken
faiss-cpu
libmagic
python-magic
openai
langchain-openai